requirements.txt
streamlit
pillow
torch
torchvision
transformers